Move modules around and improve tests
